requirements.txt
cvxopt==1.3.2
cycler==0.11.0
datasets==4.0.0
evaluate==0.4.5
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy==1.26.0
openpyxl==3.1.5
pandas==2.3.1
Pillow==9.0.1
PyYAML==5.4.1
scikit-learn==1.7.0
scipy==1.8.0
torch==2.7.1
torchmetrics==1.7.4
torchvision==0.22.1
transformers==4.53.1
